apps/mcdonalds-voice-menu/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "advanced-alchemy>=1.8.2",
    "alembic>=1.18.3",
    "asyncpg>=0.31.0",
    "azure-cognitiveservices-speech>=1.47.0",
    "litestar[standard]>=2.19.0",
    "msgspec-ext==0.4.0",
    "openai>=2.16.0",
    "pgvector>=0.4.2",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.2",
    "sqlalchemy[asyncio]>=2.0.46",
    "torch>=2.10.0",
    "uvicorn>=0.40.0",
    "deepgram-sdk>=4.0.0",
]
